feat: expose web3 signer on Keypair class
For libraries to consume this correctly there needed to be an easier way to construct the keypair with the correct signer — otherwise all nucs would need to be manually signed.

To be honest this might break the model — eg do we just need a Signer abstraction and do away with keypair? I'm punting that decision given how close we are to release.